+++ tests/unit/extract/test_simulations.py
@@ -106,7 +106,6 @@ def test_simulations_from_config_filtered_by_simulation_id(
                 "circuit": mock_circuit0,
             },
         ],
-        index=[1],
     )
     expected_df = ensure_dtypes(expected_df)
     assert isinstance(result, test_module.Simulations)
@@ -152,7 +151,6 @@ def test_simulations_from_config_without_spikes(
                 "circuit": mock_circuit0,
             },
         ],
-        index=[1],
     )
     expected_df = ensure_dtypes(expected_df)
     assert isinstance(result, test_module.Simulations)
@@ -201,7 +199,6 @@ def test_simulations_from_config_first_nonexistent(
                 "circuit": mock_circuit1,
             },
         ],
-        index=[1],
     )
     expected_df = ensure_dtypes(expected_df)
     assert isinstance(result, test_module.Simulations)
@@ -459,7 +456,6 @@ def test_simulations_from_pandas_filtered_by_simulation_id(
                 "circuit": mock_circuit0,
             },
         ],
-        index=[1],
     )
     expected_df = ensure_dtypes(expected_df)
     assert isinstance(result, test_module.Simulations)